Overall apps timeline graph implemented Individual app per-connection graph implemented Graph interval and viewsize adjustable in reasonable scales from menu

+++ src/com/googlecode/iptableslog/PacketGraphItem.java
@@ -3,15 +3,15 @@ package com.googlecode.iptableslog;
 import android.os.SystemClock;
 
 public class PacketGraphItem {
-  long timestamp;
-  long len;
+  double timestamp;
+  double len;
 
-  public PacketGraphItem(long len) {
+  public PacketGraphItem(double len) {
     this.timestamp = System.currentTimeMillis();
     this.len = len;
   }
 
-  public PacketGraphItem(long timestamp, long len) {
+  public PacketGraphItem(double timestamp, double len) {
     this.timestamp = timestamp;
     this.len = len;
   }
